+++ src/main/kotlin/com/flyn/eeg_receiver/view/Viewer.kt
@@ -2,10 +2,10 @@ package com.flyn.eeg_receiver.view
 
 import com.fazecast.jSerialComm.SerialPort
 import com.flyn.eeg_receiver.data.DataReceiver
-import com.flyn.eeg_receiver.event.SpectralEvent
+import com.flyn.eeg_receiver.event.BandEvent
 import com.flyn.eeg_receiver.event.RawWaveEvent
 import com.flyn.eeg_receiver.getLoader
-import com.flyn.eeg_receiver.listener.SpectralListener
+import com.flyn.eeg_receiver.listener.BandListener
 import com.flyn.eeg_receiver.listener.RawWaveListener
 import javafx.application.Platform
 import javafx.collections.FXCollections
@@ -21,7 +21,7 @@ import javafx.stage.FileChooser
 import javafx.util.StringConverter
 import java.util.concurrent.ConcurrentLinkedQueue
 
-class Viewer: RawWaveListener, SpectralListener {
+class Viewer: RawWaveListener, BandListener {
 
     companion object {
 
@@ -34,8 +34,8 @@ class Viewer: RawWaveListener, SpectralListener {
     private var comPort = ""
     private val rawData = XYChart.Series<Long, Int>()
     private val tempData = ConcurrentLinkedQueue<XYChart.Data<Long, Int>>()
-    private val spectralEventList = ConcurrentLinkedQueue<SpectralEvent>()
-    private val tempSpectralList = ConcurrentLinkedQueue<SpectralEvent>()
+    private val bandEventList = ConcurrentLinkedQueue<BandEvent>()
+    private val tempBandList = ConcurrentLinkedQueue<BandEvent>()
 
     @FXML
     private lateinit var rawDataChart: LineChart<Long, Int>
@@ -47,17 +47,17 @@ class Viewer: RawWaveListener, SpectralListener {
     private lateinit var connectButton: Button
     @FXML
     private lateinit var controlButton: Button
+    @FXML
+    private lateinit var saveButton: Button
 
     @FXML
     fun initialize() {
         DataReceiver.addListener(this)
         rawDataChart.data.add(rawData)
-        xAxis.lowerBound = DataReceiver.startTime.toDouble()
-        xAxis.upperBound = DataReceiver.startTime.toDouble()
         xAxis.tickLabelFormatter = object: StringConverter<Number>() {
 
             override fun toString(`object`: Number): String {
-                val currentSec = (`object`.toLong() - DataReceiver.startTime) / 1000
+                val currentSec = (`object`.toLong()) / 1000000000
                 return "$currentSec s"
             }
 
@@ -80,37 +80,39 @@ class Viewer: RawWaveListener, SpectralListener {
             isPause = false
             connectButton.text = "Connect"
             controlButton.text = "Pause"
+            controlButton.isDisable = true
+            saveButton.isDisable = true
             DataReceiver.disconnect()
         }
         else {
             connectButton.text = "Disconnect"
+            controlButton.isDisable = false
+            saveButton.isDisable = false
             DataReceiver.connect(comPort)
             rawData.data.clear()
-            xAxis.lowerBound = DataReceiver.startTime.toDouble()
-            xAxis.upperBound = DataReceiver.startTime.toDouble()
+            xAxis.lowerBound = 0.0
+            xAxis.upperBound = 0.0
         }
     }
 
     fun chartControl() {
-        if (!DataReceiver.isConnect) return
         if (isPause) {
             controlButton.text = "Pause"
             rawData.data.clear()
             rawData.data.addAll(tempData)
-            spectralEventList.clear()
-            spectralEventList.addAll(tempSpectralList)
         }
         else {
             controlButton.text = "Resume"
             tempData.clear()
             tempData.addAll(rawData.data)
-            tempSpectralList.clear()
-            tempSpectralList.addAll(spectralEventList)
+            tempBandList.clear()
+            tempBandList.addAll(bandEventList)
         }
         isPause = !isPause
     }
 
     fun saveFile() {
+        if (!isPause) chartControl()
         val fileChooser = FileChooser().apply {
             extensionFilters.add(FileChooser.ExtensionFilter("CSV files (*.csv)", "*.csv"))
         }
@@ -121,14 +123,14 @@ class Viewer: RawWaveListener, SpectralListener {
             rawData.data.forEach {
                 if (isFirst) {
                     isFirst = false
-                    appendText("${it.xValue - DataReceiver.startTime}, ${it.yValue}")
+                    appendText("${it.xValue}, ${it.yValue}")
                     return@forEach
                 }
-                appendText("\n${it.xValue - DataReceiver.startTime}, ${it.yValue}")
+                appendText("\n${it.xValue}, ${it.yValue}")
             }
-            spectralEventList.forEach {
+            tempBandList.forEach {
                 with(it) {
-                    appendText("\n${time - DataReceiver.startTime}, $delta, $theta, $lowAlpha, $highAlpha, $lowBeta, $highBeta, $lowGamma, $midGamma")
+                    appendText("\n$time, $delta, $theta, $lowAlpha, $highAlpha, $lowBeta, $highBeta, $lowGamma, $midGamma")
                 }
             }
         }
@@ -152,14 +154,9 @@ class Viewer: RawWaveListener, SpectralListener {
         }
     }
 
-    override fun onSpectralEvent(event: SpectralEvent) {
-        if (isPause) {
-            tempSpectralList.offer(event)
-            if (tempSpectralList.size > 5) tempSpectralList.poll()
-            return
-        }
-        spectralEventList.offer(event)
-        if (spectralEventList.size > 5) spectralEventList.poll()
+    override fun onSpectralEvent(event: BandEvent) {
+        bandEventList.offer(event)
+        if (bandEventList.size > 5) bandEventList.poll()
     }
 
 }